Allow thread index view to sort oldest first
This is basically the original patch by Keith Packard, adapted to the
current code base, but with one little change in thread-index-mode: the
list of threads must be reversed when newest_first is not set.

To activate it, put this in your config.yaml

    :inbox_newest_first: false

See http://article.gmane.org/gmane.mail.sup.general/2757

Original commit message:

This allows the thread index view to present oldest messages first
instead of always presenting newer messages at the top of the list.
This can be toggled using the 'o' key.

A config file option :inbox_newest_first controls whether this is
turned on by default when viewing the inbox

Refine searches inherit the order from the originating thread index.